PLANET-8246 Add video support to Timeline block
This is to provide more options for editors to illustrate the content
